autotest - fix dev_server bytes/str err
This might lead to wack-a-mole because the original b/ was not caught by
the unittest, and I am not the owner/expert on this code, so I can't
even run it locally. This specific issue I tested by repro'ing a `if str
in bytes` then handled that in a local script.

BUG=b:217780680
TEST=cq, unittest, local script

Change-Id: Ib04b454914682d484932f5f038331a6a81c7cdfc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3437887
Tested-by: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Jesse McGuire <jessemcguire@google.com>
Commit-Queue: Jesse McGuire <jessemcguire@google.com>