autotest: Remove the reset of jack plugger
We had an assumption that the default state of the jack plugger should
be plugged. However, the default state is going to be revert back to
unplug.

Having this kind of assumption in the codes is not ideal, a better
solution is to expose the plugger's state so the test may revert it
after finishes. Currently, there is no should mechanism and thus we'll
rely on the audio board's reset.

BUG=b:217986471
TEST=deploy and run test

Change-Id: If0a901610d28ae40054f5ac9457a16cdb8f1e90e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3445267
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>