Bluetooth: check connection status before the real test
There is an issue on chameleon when a DUT connects to multiple peers
with at least one emulated as an audio device, the other connections
might be dropped for a few seconds then reconnect. This is because bluez
always expects to be a central device when it accepts A2DP/AVCTP
connection and there is no way to change the expectation.

To allow EP tests to connect both a HID and an audio device, we add a
polling test to check the connection status before the real test if we
expect it can be connected.

BUG=b:210379084
TEST=run bluetooth_AdapterEPHealth.md_*

Change-Id: I38991b286b740d5824610fe86684f199903ea598
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3461130
Tested-by: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Yun-Hao Chung <howardchung@chromium.org>