Reload DBUS daemon when sideloading chrome
This is to address a problem we have been seeing sporadically in the
linux-nearby-chrome-fyi. As it turns out, sometimes when there is a
difference in dbus/chrome_dbus_services.conf between the one in
rootfs vs the one we are trying to sideload (from Chromium ToT),
it leads to errors like:
FATAL chrome: Failed to own: org.chromium.FuseBoxService

Now upon side loading chrome, Dbus daemon is restarted to ensure that
the latest policy from the currently mounted /opt/google/chrome/dbus
is used, and chrome will be authorized to take ownership of those
services.

BUG=b/219962805
TEST=test_that --board ${BOARD1} --model ${MODEL1} --companion_hosts ${DUT2} --ssh_verbosity 3 ${DUT1} tast.nearby-share-cb2cb-chrome-from-tls

Change-Id: I94e88e9a91b869d46a65c8fdae9ff6be1a0084fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3466787
Reviewed-by: David Haddock <dhaddock@chromium.org>
Tested-by: Jonathan Fan <jonfan@chromium.org>
Commit-Queue: Jonathan Fan <jonfan@chromium.org>