nbr: Refactor update_engine_tests to support NBR tests
Extracted helper method for staging the stateful archive out
out of `_restore_stateful` so that it could be used for
staging the stateful archive independent of the restore.
The new helper method also supports staging stateful archive
on a public bucket. Also extracted a method to download and
unarchive stateful archive to a directory out of
`_restore_stateful`. This enables unarchiving stateful for
other purposes other than restoring which is needed by
NBR tests.

BUG=b:203691549
TEST=test_that -b ${B} ${H} nbr_EndToEndTest.basic --args="job_repo_url='${devserver}' running_at_desk=True"

Change-Id: Idfc9eeaf84015f91002d96ae0d89abd9124fccb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3475699
Tested-by: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: Henry Barnor <hbarnor@chromium.org>