Bluetooth: Only connect one peer at a time in EP MD tests
The workaround crrev/c/3461130 is trying to handle issue b/210379084,
but it might not be enough since the peer can be disconnected after we
check the connection status.

This workaround is to make sure we only connect one device at a time if
the issue b/210379084 can happen.

BUG=b:210379084
TEST=run bluetooth_AdapterEPHealth.md_* on ezkinil in skylab

Change-Id: I0fcc5253fe5ea844e8b263d264ccb3cf8ff56e9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3490549
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Yun-Hao Chung <howardchung@chromium.org>
Tested-by: Yun-Hao Chung <howardchung@chromium.org>