FAFT: Utilize the return values of restore_firmware and restore_kernel
Since we already support returning a bool value denoting if the firmware
was changed in CL:1723427, refactor the codes and remove one call in
_reset_client() to utilize its return value.

Also add the return value of restore_kernel, which returns if the kernel
or MiniOS is changed. The logic is the same as the firmware side.

BUG=none
TEST=none

Change-Id: Ie1a0cb318a97fd489769e81ba0bef1890bb8f5a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3498133
Auto-Submit: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Tested-by: Hsuan Ting Chen <roccochen@chromium.org>
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>