autotest - fix crash causing hooks to silently end.
Often a symlink source will be missing for whatever reason. this causes
either the pre/post test hook to crash when it goes to look for that
file. This crash stops the hook, so any non-collected logs will remain
as such. This can cause incomplete and often confusing logs.

So instead, if a file is missing, we will simply skip that file and
continue with the hooks.

BUG=b:195527581
TEST=locally made the hook crash, then applied the fix and all other
files were there.

Change-Id: I61b642d6ee92b92d3a4e51fa0866d5470b42af6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3498897
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Jesse McGuire <jessemcguire@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Tested-by: Derek Beckett <dbeckett@chromium.org>