audio_CrasAec: Wait for AEC stream ready by polling
Previously in https://crrev.com/c/3308233 we increased the wait
time to 0.3 seconds but that is still not enough. On some platforms
we see it take quite a long time for AEC stream to be created.

This change addresses the issue by changing to poll and wait
method.

BUG=b:205921665
TEST=test_that <dut-ip> audio_CrasAec
 emerge-atlas autotest-tests-audio

Change-Id: I9d945e23fa7979630e7316810ce0b1fe35c30ab0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3505967
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>