autoupdate: Handle potential network screen skip
CL:3452268 landed in M100 and caused test failures in
autoupdate_ForcedOOBEUpdate on the Chrome OS CQ. Therefore, it was
reverted in CL:3498329.

This CL modifies autoupdate_StartOOBEUpdate.py to handle cases in which
the network screens should be skipped. In these cases, the test won't
wait for the screen to be visible nor click "next".

BUG=b:221254005, chromium:1256890
TEST=test_that -b ${BOARD} chromebook autoupdate_StartOOBEUpdate

Change-Id: I218f6d321d342c32c6d5eadb238d7c558bc4ee7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3509829
Tested-by: Felipe Andrade <fsandrade@chromium.org>
Reviewed-by: Roman Aleksandrov <raleksandrov@google.com>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Felipe Andrade <fsandrade@chromium.org>