bluetooth: Support classic pairing test with Floss
Add support for pairing test in classic health. This adds pairing
support to the Floss facade and updates the FlossAdapterClient with all
the needed apis.

There's also a few fixes and updates to the Floss utility classes for
the PropertySet and the glib_call decorator.

BUG=b:216348040
TEST=test_that bluetooth_AdapterCLHealth.all_floss

Change-Id: Ic8d99513423793b6700c2faf4ef0b50ee1008d42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3518309
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@google.com>