bluetooth: Fix smart unlock test
In exchange_messages, mainloop.run() was never exiting if the property
was updated and the timer source was removed. This doesn't always happen
and seems like a race/bug in GLib itself. To work around this, we don't
remove the timer source in the property changed signal and just mark an
early exit for the timeout to cleanly exit instead.

BUG=b:222364364
TEST=Stress test bluetooth_AdapterLEBetterTogether.smart_unlock_test

Change-Id: I5fa19695843f2592f6d88acb578b0e21798ed1bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3526057
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Auto-Submit: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Reviewed-by: Michael Sun <michaelfsun@google.com>
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@google.com>