update_engine_test: Fix copy payload bug that added subdirectory
http://crrev/c/3475699 introduced a bug that caused payloads to be
copied into a subdirectory. The  src-url given to gsutil always
included a globbing pattern at the end which changes the meaning
of the dest-url to be a directory when the glob matches more than
one file.

This change resolves the issue by adding a parameter to control the
globbing. It defaults the parameter to the previous behavior of always
enable globbing and ensures that the destination url is only changed
from the public bucket when a destination filename is provided.

BUG=b:226982187
TEST=Ensure payload is not already in public bucket: `gsutil ls 'gs://${PUBLIC_BUCKET}/${VERSION}*'`
TEST=test_that -b ${B} ${H} autoupdate_Basic.full --args="job_repo_url='${devserver}/${VERSION}' running_at_desk=True"
TEST=test_that -b ${B} ${H} nbr_EndToEndTest.basic --args="job_repo_url='${devserver}/${VERSION}' running_at_desk=True"

Change-Id: Ib94d77031c6d954e8e36e184203a704f02f17637
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3555576
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Henry Barnor <hbarnor@chromium.org>
Commit-Queue: Henry Barnor <hbarnor@chromium.org>