tast: Only try to get all devservers on satlab
Apparently passing prefer_local_devserver=False to
get_available_devservers only works in mainlab (where it is not needed)
and satlab (where is it required), but not in moblab or chroots.

Change the logic from "not moblab" to "is satlab".

BUG=b:207133139,b:213964936
TEST=test_that tast.firmware-usb, verified no devservers passed to tast.

Change-Id: I4cdf7389e44d49f58d69c4331394ae3a34220d78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3558210
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>