sys_power: waive the emmc error from firmware for brya/primus
Some skus of the brya/primus don't include the emmc and the firmware
will prompt the error message during the system boot up. On the other
hand, the power_SuspendStress will check the error in the firmware log
and which makes the power_SuspendStress test failed. In order to fix
this problem, we can waive the error message from the test for the
brya/primus.
Note: this waiver won't impact the suspend test to detect the real
error because we don't log the S0ix log in the firmware log now.

BUG=b:195336611
TEST=test_that f:.*power_SuspendStress/control.bareFSI

Change-Id: Ie11bb9305e96a2a3af02700baf4e10fdbb9dabc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3559874
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>