autotest: Rename hardware_TrimIntegrity control files
hardware_TrimIntegrity has been migrated to python 3, but the control
file names start with _, which is indicating that they should be deleted.
We use this test in storage qualification, so we do not want it to be removed.

Also, hardware_StorageFio does not allow testing the entire root device, so fix
hardware_TrimIntegrity to not use filesize=0 for hardware_StorageFio call.

BUG=b:201238564
TEST=test_that hardware_TrimIntegrity

Cq-Depend: chromium:3584682
Change-Id: I923d696bf974fd4715dea30115a3d25b01938891
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3579557
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Alexis Savery <asavery@chromium.org>
Tested-by: Alexis Savery <asavery@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>