Add m2n variant of autoupdate_EndToEndTest
Add an m2n variant of autoupdate_EndToEndTest. This variant will run
outside of paygen and use SSP. This lets us make sure the autotest-main
version of this test works fine in the lab before we turn on SSP for the
paygen tests. This variant will also be added to the CQ once stabilized,
to catch CLs that break this test early and minimize disruptions to the
release process.

The changes to the test and auto_updater* files are to handle the
different payload source that's used in the typical AU tests. For
paygen, we test signed payloads (the ones from gs://chromeos-releases/).
The rest of the AU test framework uses the payloads from
gs://chromeos-image-archive/.

BUG=b:228100799
TEST=test_that <DUT> autoupdate_EndToEndTest.m2n.full --args="build=R102-14692.0.0"

Change-Id: I2811596e4abb0bcb4a2739284d3cbd55999599b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3587628
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>