network_WiFi_BluetoothStreamPerf: Backport a three-way Perf setup
Utilize the pcap device as an iPerf/Netperf endpoint in throughput
testing. This setup utilizes some hard-coded values for ethernet
interface names and ip addresses for the LAN ports on the pcap and
router Gale devices. This is necessary because the router device is
not configured with the daemons to handle the assignment automatically.
This also means that some ip routes need to be set up manually on the
devices to allow traffic to move as needed.

Specific arguments for the LAN interface device names and ip addresses
may be provided at the command line. If not provided, default values
will be used instead.

This is a backport of the solution implemented in
https://chromium-review.googlesource.com/3108426

Common functionality shared by network_WiFi_Perf.py and
network_WiFi_BluetoothStreamPerf.py were moved to the base class
WiFiCellPerfTestBase

BUG=b:224781062,b:184755043,b:184758784,b:184764025,b:225284551
TEST=test_that --args "btpeer_host_list=${RASPBERRY_IP} \
        router_addr=${AP_IP} pcap_addr=${PCAP_IP}" ${DUT_IP} \
        suite:bluetooth_wifi_coex
TEST=test_that --args "router_addr=${AP_IP} pcap_addr=${PCAP_IP}" \
        ${DUT_IP} suite:wifi_perf

Change-Id: Ib881e42a43ac9f8bcdaf215baa477e20a2b3f04b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3582929
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Tested-by: Jacek Klimkowicz <jakl@semihalf.com>
Commit-Queue: Jacek Klimkowicz <jakl@semihalf.com>