WiFiCellPerfTestBase: Apply governor settings for pcap host
If 'governor' is provided as CLI option it will be also applied to
the pcap host which is the packets generator under three-way-setup.

If the default governor name is different on router and pcap host,
the name 'default' will be logged, otherwise the real governor
settings will be used.

BUG=b:224781062,b:184755043,b:184758784,b:184764025,b:225284551
TEST=test_that --args "btpeer_host_list=${RASPBERRY_IP} \
        router_addr=${AP_IP} pcap_addr=${PCAP_IP}" ${DUT_IP} \
        suite:bluetooth_wifi_coex
TEST=test_that --args "router_addr=${AP_IP} pcap_addr=${PCAP_IP}" \
        ${DUT_IP} suite:wifi_perf

Change-Id: I155a805f0aac924bee03bf6bee2061111f1e1715
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3834384
Commit-Queue: Jacek Klimkowicz <jakl@semihalf.com>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>