power_BrightnessResetAfterReboot: fix range for ALS
If the ambient light is near 0 lux, the current
brightness range that uses lux * 2 and lux / 2
is practiacally the same level.

This can make the test fail with rounding error,
so this CL adds an absolute  offset for the range
for device with ALS to deal with that.

BUG=b:224767649
TEST=Ran the test

Change-Id: I5fd11fcbbe898cf0904c90e00dfb43606d0a10d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3598433
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>