bluetooth: let a dut capture audio until it is stopped
The test_dut_to_start_capturing_audio_subprocess() was used to enable
the HFP profile for a pre-defined duration. The follow up methods of
playing and recording on the dut and the device may be slow such that
the recorded stream is not as long as the expected duration. This
patch sets duration to None so that the capturing audio subprocess
can last until it is stopped explicitly.

BUG=b:199406747, b:185942785
TEST=The failure message "failed to trim recorded file" is caused by
a few issues. With this patch, such failures are reduced.
Also all A2DP and HFP tests in bluetooth_AdapterAUHealth should pass.

Change-Id: I2a7721dc66b5cad975351137e25f52ba1436fea1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3607544
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>