pvs: implement pvs_Sequence with example .Example
the power_qual and thermal_qual requirements both instrument series of
tests that guarantee DUT state, and may need to record performance data
and results from multiple tests in the series. This sequence
standardizes a mechanism to capture a multi-test series in a control
file, and surface those results as first-class test results in the logs.

BUG=b:229991584
TEST=test_that $DUT pvs_Sequence.Example and upload to CPCon

Change-Id: I710546c567833100eef118307bec5ee8ca597adc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3625576
Commit-Queue: Parker Holloway <jparkerh@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Tested-by: Parker Holloway <jparkerh@google.com>