faft-cr50: add test to verify setting the hash with ap ro v1
Add a test that verifies GSC can set the hash when the board id type is
erased.

Verify the board id can be set and cleared when the board id is
completely erased.

Verify the board id can be set and cleared when the board id flags are
set, but the type is erased.

Verify the board id can't be set or cleared after the board id type has
been set.

BUG=b:218705748
TEST=run on hatch

Change-Id: Id1d6f8ac5cc956618ca44bac60fbebc7380f4c53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3646282
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>
Auto-Submit: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>