platform_ServoPowerStateController: check power state in assert_dut_off
Currently, the assert_dut_off() function only verifies that the dut is
not pingable to check that it is off.  However, on the herobrine
board, when this check occurs, the dut is actually booted into the
firmware screens, so it is registered as off even though it is not.
Adding an additional check for the power state to make sure that the
dut is actually off.

BUG=b:215789278,b:230663570
TEST=Make sure that platform_ServoPowerStateController.USBPluggedin
     fails on the herobrine board because currently power_state:rec
     followed by power_state:off results in the board booting into the
     OS when it shouldn't (due to a bug).

Change-Id: I1123512fbe8e67238ef7e0b637a6426fa6bc2df7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3636954
Reviewed-by: Bob Moragues <moragues@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>
Commit-Queue: Shelley Chen <shchen@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>