bluetooth: Move identify_platform_failure_reasons to quick_test_init.
After this patch, the batch wrapper includes no server test-specific
code and can be smoothly moved to common_lib.

We always run quick_test_init() before running batches, so the
behavior should remain the same on most tests.
The only affected test batch is QuickHealth which runs multiple
batches, so before this patch the platform failure reason is set for
multiple times but that is actually not necessary.

BUG=b:232038261
TEST=bluetooth_AdapterAdvHealth.adv_peer_test
Print out the COMMON_FAILURES and it is correct.

Change-Id: I089c9e217d4f3931a0d51e006ba245fa34488b1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3640062
Tested-by: Hsin-chen Chuang <chharry@google.com>
Commit-Queue: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>