cheets_CTS_R: Temporary workaround to tweak CPUfrep governor.
The current Vulkan dEQP on ARM ARCVM takes more than
16 hours to finish. To fit them in the job time limit,
we boost the CPU to the performance mode.

This will be removed once the Vulkan impl speeded up,
or we could come up with a nice way to split dEQP jobs
to smaller pieces.

BUG=b:228879007
TEST=cheets_CTS_R.internal.arm.CtsDeqp.32 on trogdor-arc-r

Change-Id: I674f2d8824a3e2dc3ec49ab542ef59e537ca3b3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3640443
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Shao-Chuan Lee <shaochuan@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Yiwei Zhang <zzyiwei@chromium.org>