autotest: Correctly handle timeout in iperf warm up
The iperf_runner server setup runs an iperf command for 10 seconds to
ensure that the server has enough time to come online. The result of
this iperf command is supposed to be completely ignored, except for
some debug logging. However, when ignore_timeout is passed to the run()
function, SSH command timeouts result in None being returned from the
run() function.

(Source: http://cs/chromeos_public/src/platform/tauto/src/autotest_lib/server/hosts/ssh_host.py?l=272)

This means that when we check result.exit_status in the iperf server
setup, we get an exception if a timeout occurred. This change adds a
check to handle a None result value correctly, so we don't hit this
exception. The end result is that we should move on to the main perf
test, which has the logic to handle incorrect connectivity robustly.

BUG=b:214403123
TEST=Adjust timeout duration in the code to artificially trigger a
timeout every time we run the test. Without this CL, we hit the
"NoneType" failure case. With this CL, we correctly move on to the main
testing section, which handles the failure more robustly.

Change-Id: Ie10824d8d4b19d7174bf36dca996543f8e079493
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3642500
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>