Use non-empty credentials in autoupdate tests
The default telemetry login uses an empty password, which will soon be
changed since empty passwords will be disallowed. Changing the default
password on the telemetry side breaks the AU tests, since they will use
the old (empty) password to sign in while on the source version, and the
new (non-empty) password to sign in on the target version.

To fix this, add arguments to login_LoginSuccess to specify the
credentials, and modify the AU server tests to explicitly pass in
consistent credentials. Even on m2n tests and autoupdate_EndToEndTest,
which provision the DUT to older milestones, the actual client test code
should come from the target build. This is because autotest downloads
the autotest artifacts from the chromeos-image-archive bucket using the
build in the job_repo_url, and pushes those to the DUT to run client
tests.

BUG=chromium:1325411
TEST=autoupdate_Basic.m2n.full, autoupdate_DataPreserved.m2n.full,
autoupdate_EndToEndTest.m2n.full

Change-Id: Ic919d4f8fd5edf4a26e6a27624698ad5f4feb98a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3642508
Reviewed-by: David Haddock <dhaddock@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>