nbr: Make MiniOS tests more resilient to failured recovery
MiniOS tests can have a failure condition that leaves the current active
partition unbootable. To avoid this having an effect on other tests in a
test suite, we provision the current running version of ChromeOS on the
inactive partition to ensure that failure cases leave the DUT running
with the expected version of ChromeOS for tests in the same suite.

BUG=b:223643256
TEST=test_that -b ${B} ${H} suite:nbr --args="job_repo_url='${repo_url}'"

Change-Id: Ib59cf0273589f5c0f84ec3033f7733c9aaa7b53f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3658244
Tested-by: Henry Barnor <hbarnor@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: Henry Barnor <hbarnor@chromium.org>