faft: dm-verity / FW rootfs corruption test
The test dumps verity hash tags to a persistent location, then corrupts
those present on actual rootfs to trigger panic and custom error
behaviour which marks the kernel unbootable. The test then ensures FW
correctly failed over to kernel B. At the end the test restores the
verity hash tags on the rootfs and boot args of kernel A and verifies
that kernel A is bootable again.

BUG=b:149112468
TEST=test_that --board=cherry 192.168.1.12 kernel_VerityCorruptRootfsA.dev

Change-Id: I2624f33d18c0c1767767813047ba7991e8d7e5e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3682905
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>