Replace chromite usage in au_e2e with standard update_engine_test flow
The frozen_chromite usage in au_e2e is blocking the SSP transition for
this test, which is important for improving this test's stability and
reliability. All other AU tests manage the nebraska update flow directly
from update_engine_test/update_engine_util, and there don't appear to be
any network-related issues (which is the main thing the chromite way is
meant for). Switching away from frozen_chromite will unblock the SSP
conversion and make this test use the same update flow as all other
tests, so let's try it out. Once we confirm that the SSP m2n variant is
running well in the lab, I will schedule this new version of the test in
paygen to run alongside the existing autotest-drone version so we can
compare the results before fully moving over to SSP.

BUG=b:228100799
TEST=autoupdate_EndToEndTest/.ssp/.m2n.full

Change-Id: I401b10735efeacdd69cd82ea3c5f0978522d4881
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3687634
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Kyle Shimabukuro <kyleshima@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>