power_VideoCall: ensure system services are all down
power_VideoCall uses the chrome browser, which was restarting the powerd
service. This affects the test results  (by dimming the screen, etc),
and needed to be disabled in order to ensure consistency across devices
and test cases. Call the system service shutdown utility after launching
the browser and pass "no firmware update app" option to the instance of
the browser.

BUG=b:234622928
TEST=run power_VideoCall, check "status powerd" mid-test over ssh

Change-Id: Id449039e0b2e2fe2897818d3c3325116011cd1fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3687635
Commit-Queue: Parker Holloway <jparkerh@google.com>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Parker Holloway <jparkerh@google.com>