upload_results: unique job id with current timestamp
The upload_results tool previously used the timestamp from the start of
the job to tag the job id for the upload. This created issues with
re-uploaded and generated results, as both could have the possibility
for collisions in job-id (and would be rejected by the pipeline). This
CL swaps over to use the current time in milliseconds to tag the job,
resolving both of those issues.

BUG=b:216854680
TEST=upload stub_Pass result twice, with the -f flag

Change-Id: I0119335115504bfed42915c9e503cf2870862734
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3687637
Tested-by: Parker Holloway <jparkerh@google.com>
Reviewed-by: Svitlana Soltysyak <ssoltysyak@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Parker Holloway <jparkerh@google.com>