UploadResults: update job structure to include mandatory fields
The job structure was previously discarding "suite", "afe_job_id, and
other fields when serializing and deserializing job protobufs. Modify
the constructor for the job to include these fields as optional
arguments, so that when we load/unload job.serialize we are picking up
the mandatory fields. This keeps us from having to re-parse all of these
fields each time (and in some cases when the necessary information is
unavailable).

BUG=None
TEST=uploading using the PVS tool, and via command line

Change-Id: Ibfe414ff2c4a997c36c86de90ba64eb549738b8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3692931
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Parker Holloway <jparkerh@google.com>
Tested-by: Parker Holloway <jparkerh@google.com>