power_WakeSources: Use the platform name to query the faft_config
The power_WakeSources is using the get_platform() from the cros_host.py
to get the platform name for the faft_config. However, the get_platform
returns the model name instead of the platform name. With the model name
being used, we need to add every model's config to the fw-testing-configs
but we should use the platform config instead. This patch changes the
faft_config by reading the platform name via the RPCProxy.

BUG=b:230398487
TEST=test_that -b brask ${IP} power_WakeSources

Change-Id: I2c97b9ac01d1a2941550c4cd59252dc78f7109dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3699501
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Tested-by: Ariel Fang <ariel_fang@wistron.corp-partner.google.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>