power_SuspendStress: Fail instead of warn on kernel/timeout failures
This makes power_SuspendStress fail instead of throw a warning when the
failures are only kernel and timeout errors. This is hiding actual
issues, possibly even with firmware since a kernel error could actually
be due to firmware. Since the minimum suspend time is 15 seconds, we
should not be getting timeouts unless there's an actual issue we should
fix.

BUG=b:233962395
TEST=power_SuspendStress.bareFSI

Change-Id: I273fcfc83b742fbd71b9db7971b149767ea39a3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3706072
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>