sys_power: Waive Ti50 firmware version error on nissa
Some nissa devices are still using Ti50 0.0.12, which doesn't support
reading the firmware version. This results in the following coreboot
error:

[ERROR]  Reading Ti50 firmware version is not supported

This doesn't actually cause any issues because the firmware version is
only read in order to print it to the console. So waive this error on
nissa to allow suspend/resume tests to pass.

This workaround will be removed once all Ti50 stocks are updated to
0.0.15 or higher, tracked in b/236911319.

BUG=b:233972322
TEST=power_Resume no longer fails with "FirmwareError(1): [ERROR]
Reading Ti50 firmware version is not supported" on nivviks (there are
still other errors which are being fixed elsewhere).

Change-Id: Ia92a995b0766e4235153e4774aa8f66926b3ff20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3703901
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Auto-Submit: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Kangheui Won <khwon@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>