tradefed_test: Check gsutil hash before downloading
Use gsutil hash to find the hash of the file, to determine whether a
redownload is necessary.
For gs URLs, instead of hashing the URL, the hash of the file is used to
determine whether it has been downloaded (cached).

BUG=b:236197850, b:236092837
TEST=./utils/unittest_suite.py
server.cros.tradefed.tradefed_test_unittest

Change-Id: I98fd17ed4f93b68c4e3e0cc8f0487b7039e47492
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3716316
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Reviewed-by: Shao-Chuan Lee <shaochuan@chromium.org>
Commit-Queue: Rintaro Kuroiwa <rkuroiwa@google.com>
Tested-by: Rintaro Kuroiwa <rkuroiwa@google.com>