Reenabling some more Chameleon HDMI tests.
Now that we have Chameleon V3 in the lab and the basics working,
its time to enable more tests to gather more information & feedb-
ack.

Note about 'TEST=None' below: display_LidCloseOpen test is flagging
an error because it is not listed in the ebuild file to be picked
up as a test. However, the ebuild file is in a different part of
the source tree and will be updated separately. For now, I am
uploading this change to gerrit with '--no-verify'.

BUG=b:234759299
TEST=None

Change-Id: I1c7df2e3431d96f88de1fc3bf6af0817c8262cd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3717484
Tested-by: Vinayak Suley <vsuley@chromium.org>
Auto-Submit: Vinayak Suley <vsuley@chromium.org>
Commit-Queue: Alexandru Stan <amstan@chromium.org>
Reviewed-by: Alexandru Stan <amstan@chromium.org>