autotest: Hide nudges for graphics_HwOverlays
The light/dark mode UI experiment in Ash adds a nudge to the bottom
left of the screen which can overlap [1] with the content in the
hardware overlays tests. Since there is a backdrop filter on the
nudge (background blur), the content behind the nudge cannot be
promoted to an overlay, and subsequently, the tests fail.

This is a problem in Tast too, so the --ash-no-nudges flag was added
to hide the nudges during tests [2]. This CL updates our legacy
autotest test to use the same flag.

[1] https://i.imgur.com/SZ6P3ra.jpg
[2] https://chromium-review.googlesource.com/c/chromium/src/+/3727126

Bug: b:227956160
Change-Id: Ifab6a94f76c9c97a462779256ebcb3f444619c3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3738880
Tested-by: Brian Ho <hob@chromium.org>
Commit-Queue: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>