power_BatteryDrain: flag for use_design_charge
Add a new flag to the power_BatteryDrain utility/test to enable draining
to a percentage derived from the design charge_full rather than the
charge_full calculated by the device. This is necessary to align with
the battery percentages in the rest of the power test content.

BUG=b:235083509
TEST=local run of power_BatteryDrain, followed by power_BatteryCharge,
no additional elapsed time due to mismatched percentages.

Change-Id: I35576fc4648d31b919a281072566afdd0a3ecd10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3756167
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Parker Holloway <jparkerh@google.com>
Tested-by: Parker Holloway <jparkerh@google.com>