power: add a power_suspend_delay module
This patch adds a new power_suspend_delay module so that tests can
use it to enforce a specified delay before suspend.

The motivation behind this module is that some Bluetooth activities
during the pre-suspend phase waiting for suspend delays might generate
undesired events which prevent the system from entering the suspend
state. It is a regression that should be captured in tests. With this
module, it becomes easier for Bluetooth tests to capture this kind of
regressions by setting a longer suspend delay interval.

This module is placed in the power directory so that it may be useful
for other components' tests.

BUG=b:237256795
TEST=python power_suspend_delay.py. The output looks like

2022-07-13 13:47:21,560 suspend_delay started
2022-07-13 13:47:21,562 RegisterSuspendDelay: delay_id 1863188537
                        timeout 10000 us
2022-07-13 13:47:21,563 suspend started
2022-07-13 13:47:21,572 Running '/usr/bin/powerd_dbus_suspend --delay=0
                        --timeout=30 --wakeup_count=13874
                        --wakeup_timeout=30'
2022-07-13 13:47:21,585 SuspendImminent signal
2022-07-13 13:47:29,587 HandleSuspendReadiness: suspend_id 1863188523
                        delay 8 seconds
2022-07-13 13:47:52,694 suspend ended
2022-07-13 13:47:52,695 SuspendDone signal: duration 31.111 seconds
                        wakeup_type OTHER
2022-07-13 13:47:52,917 UnregisterSuspendDelay: delay_id 1863188537
2022-07-13 13:47:52,918 cleanup
2022-07-13 13:47:52,918 suspend_delay done

Change-Id: I8b3ee28dbb16c6a0958f23fafe736ebb178ed3ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3787938
Reviewed-by: Ying Hsu <yinghsu@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>